uniendo algunos componentes
