+++ src/arrayContent.js
@@ -1,0 +1,528 @@
+const content = [
+  {
+    id: 0,
+    lugar: "lugar 1",
+    provincia: "Tucumán",
+    categoria: "montaña",
+    descripcion: "descripción del lugar 1",
+    liked: false,
+    img: {
+      img1: "https://www.welcomeargentina.com/rio-negro/imagenes/rio-negro.jpg",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: false,
+    destacado: true,
+    user: "Chapulin Colorado",
+    userProfile: "https://pbs.twimg.com/media/EVcjT-JXQAIq0Yy.jpg",
+  },
+  {
+    id: 1,
+    lugar: "lugar 2",
+    provincia: "Tucumán",
+    categoria: "selva",
+    descripcion: "descripción del lugar 2",
+    img: {
+      img1: "https://www.viajes.com/wp-content/uploads/destinos-tc2/misiones-argentina.jpg",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: false,
+    destacado: false,
+    user: "Don Ramon",
+    userProfile:
+      "https://portal.andina.pe/EDPfotografia/Thumbnail/2013/09/02/000218071W.jpg",
+  },
+  {
+    id: 2,
+    lugar: "lugar 3",
+    provincia: "Bs As",
+    categoria: "montaña",
+    descripcion: "descripción del lugar 3",
+    img: {
+      img1: "https://vivo247.com/wp-content/uploads/2020/10/salta-sello-viaje.jpg",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: true,
+    destacado: false,
+    user: "El Zorro",
+    userProfile:
+      "https://www.lanacion.com.ar/resizer/wWlTepZ2pHuoyXZF3U7RR7C1iic=/309x206/smart/filters:format(webp):quality(80)/cloudfront-us-east-1.images.arcpublishing.com/lanacionar/MGYRO7LAGVBHBMR52AN5EKPQZY.jpg",
+  },
+  {
+    id: 3,
+    lugar: "lugar 4",
+    provincia: "Bs as",
+    categoria: "Playa",
+    descripcion: "descripción del lugar 4",
+    img: {
+      img1: "https://www.clarin.com/img/2019/01/10/jgGG_BkgV_720x0__1.jpg",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: false,
+    destacado: true,
+    user: "El Zorro",
+    userProfile:
+      "https://www.lanacion.com.ar/resizer/wWlTepZ2pHuoyXZF3U7RR7C1iic=/309x206/smart/filters:format(webp):quality(80)/cloudfront-us-east-1.images.arcpublishing.com/lanacionar/MGYRO7LAGVBHBMR52AN5EKPQZY.jpg",
+  },
+  {
+    id: 4,
+    lugar: "lugar 5",
+    provincia: "Catamarca",
+    categoria: "montaña",
+    descripcion: "descripción del lugar 5",
+    img: {
+      img1: "https://volemos.nyc3.cdn.digitaloceanspaces.com/blog/wp-content/uploads/2020/01/ruta-de-los-6miles.jpg",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: false,
+    destacado: false,
+    user: "El Zorro",
+    userProfile:
+      "https://www.lanacion.com.ar/resizer/wWlTepZ2pHuoyXZF3U7RR7C1iic=/309x206/smart/filters:format(webp):quality(80)/cloudfront-us-east-1.images.arcpublishing.com/lanacionar/MGYRO7LAGVBHBMR52AN5EKPQZY.jpg",
+  },
+  {
+    id: 5,
+    lugar: "lugar 6",
+    provincia: "Tucumán",
+    categoria: "Montaña",
+    descripcion: "descripción del lugar 6",
+    img: {
+      img1: "https://periodicoparatodos.com.ar/wp-content/uploads/2022/01/1-800x445.jpg",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: false,
+    destacado: false,
+    user: "Don Ramon",
+    userProfile:
+      "https://portal.andina.pe/EDPfotografia/Thumbnail/2013/09/02/000218071W.jpg",
+  },
+  {
+    id: 6,
+    lugar: "lugar 7",
+    provincia: "Misiones",
+    categoria: "Cataratas",
+    descripcion: "descripción del lugar 7",
+    img: {
+      img1: "https://www.expreso.info/files/2019-12/Iguazu_Catarata.jpg",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: false,
+    destacado: false,
+    user: "Chapulin Colorado",
+    userProfile: "https://pbs.twimg.com/media/EVcjT-JXQAIq0Yy.jpg",
+  },
+  {
+    id: 7,
+    lugar: "lugar 8",
+    provincia: "Jujuy",
+    categoria: "Llanura",
+    descripcion: "descripción del lugar 8",
+    img: {
+      img1: "https://dynamic-media-cdn.tripadvisor.com/media/photo-o/15/92/b9/5f/salinas-grandes-es-la.jpg?w=500&h=300&s=1",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: true,
+    destacado: true,
+    user: "El Zorro",
+    userProfile:
+      "https://www.lanacion.com.ar/resizer/wWlTepZ2pHuoyXZF3U7RR7C1iic=/309x206/smart/filters:format(webp):quality(80)/cloudfront-us-east-1.images.arcpublishing.com/lanacionar/MGYRO7LAGVBHBMR52AN5EKPQZY.jpg",
+  },
+  {
+    id: 8,
+    lugar: "lugar 9",
+    provincia: "Chubut",
+    categoria: "Montaña",
+    descripcion: "descripción del lugar 9",
+    img: {
+      img1: "https://cdn-sp.radionacional.com.ar/wp-content/uploads/2020/01/Simoca.jpg",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: false,
+    destacado: true,
+    user: "Don Ramon",
+    userProfile:
+      "https://portal.andina.pe/EDPfotografia/Thumbnail/2013/09/02/000218071W.jpg",
+  },
+  {
+    id: 9,
+    lugar: "lugar 10",
+    provincia: "Mendoza",
+    categoria: "Campo",
+    descripcion: "descripción del lugar 10",
+    img: {
+      img1: "https://dynamic-media-cdn.tripadvisor.com/media/photo-o/0e/8f/87/f9/vista-del-rio-futaleufu.jpg?w=700&h=500&s=1",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: true,
+    destacado: false,
+    user: "El Zorro",
+    userProfile:
+      "https://www.lanacion.com.ar/resizer/wWlTepZ2pHuoyXZF3U7RR7C1iic=/309x206/smart/filters:format(webp):quality(80)/cloudfront-us-east-1.images.arcpublishing.com/lanacionar/MGYRO7LAGVBHBMR52AN5EKPQZY.jpg",
+  },
+  {
+    id: 10,
+    lugar: "lugar 11",
+    provincia: "Mendoza",
+    categoria: "Montaña",
+    descripcion: "descripción del lugar 11",
+    img: {
+      img1: "https://www.clarin.com/img/2019/11/06/a8GQ9c_x_1200x630__1.jpg",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: false,
+    destacado: true,
+    user: "Don Ramon",
+    userProfile:
+      "https://portal.andina.pe/EDPfotografia/Thumbnail/2013/09/02/000218071W.jpg",
+  },
+  {
+    id: 11,
+    lugar: "lugar 12",
+    provincia: "Bs As",
+    categoria: "Llamativo",
+    descripcion: "descripción del lugar 12",
+    img: {
+      img1: "https://www.argentinaturismo.com.ar/mendoza/img/img-pozo-de-las-animas.jpg",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: true,
+    destacado: false,
+    user: "Chapulin Colorado",
+    userProfile: "https://pbs.twimg.com/media/EVcjT-JXQAIq0Yy.jpg",
+  },
+  {
+    id: 12,
+    lugar: "lugar 13",
+    provincia: "La Pampa",
+    categoria: "Playa",
+    descripcion: "descripción del lugar 13",
+    img: {
+      img1: "https://www.passagenspromo.com.br/blog/wp-content/uploads/2020/04/turismo-em-buenos-aires-floralis-generica.jpg",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: false,
+    destacado: true,
+    user: "Don Ramon",
+    userProfile:
+      "https://portal.andina.pe/EDPfotografia/Thumbnail/2013/09/02/000218071W.jpg",
+  },
+  {
+    id: 13,
+    lugar: "lugar 14",
+    provincia: "La Rioja",
+    categoria: "Montaña",
+    descripcion: "descripción del lugar 14",
+    img: {
+      img1: "https://media.viajando.travel/p/87638b222a6a67c84c54773645664f6b/adjuntos/236/imagenes/000/470/0000470375/la-rioja-pn-talampayajpg.jpg?2022-04-04-17-03-37",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: false,
+    destacado: true,
+    user: "Chapulin Colorado",
+    userProfile: "https://pbs.twimg.com/media/EVcjT-JXQAIq0Yy.jpg",
+  },
+  {
+    id: 14,
+    lugar: "lugar 15",
+    provincia: "Bs As",
+    categoria: "Ciudad",
+    descripcion: "descripción del lugar 15",
+    img: {
+      img1: "https://www.cronista.com/files/image/459/459505/62700a0f1706f.jpg",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: true,
+    destacado: false,
+    user: "Don Ramon",
+    userProfile:
+      "https://portal.andina.pe/EDPfotografia/Thumbnail/2013/09/02/000218071W.jpg",
+  },
+  {
+    id: 15,
+    lugar: "lugar 16",
+    provincia: "Tucumán",
+    categoria: "Rural",
+    descripcion: "descripción del lugar 16",
+    img: {
+      img1: "https://cdn-sp.radionacional.com.ar/wp-content/uploads/2020/01/Simoca.jpg",
+      img2: "https://upload.wikimedia.org/wikipedia/commons/2/22/Tucuman_Simoca_Plaza.JPG",
+      img3: "https://mapio.net/images-p/42423331.jpg",
+      img4: "https://media.elpatagonico.com/p/ee309b8bcd5453f7af74287b020b7f4b/adjuntos/193/imagenes/010/199/0010199640/1200x675/smart/simoca2jpg.jpg",
+      img5: "https://www.tucumanturismo.gob.ar/carga/image/festival%20sulky.jpg",
+    },
+    servicios: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    tips: [
+      "Llevar agua para mantenerse hidratados y evitar apunamiento.",
+      "Podes elegir la opción de agregar el adicional al fabuloso cerro de los 14 colores. En este caso, podes comprar el tour completo a Serranías del Hornocal.",
+      " Los restaurantes para los almuerzos en los destinos son a elección de los pasajeros. Las comidas no están incluidas.",
+      "Las excursiones pueden brindarse en buses de hasta 45 personas, en ese caso se coordinara punto de encuentro.",
+      "La salida para las excursiones puede ser con búsqueda por su hotel, en distintos puntos de encuentro a consultar o desde la oficina del operador local en pleno centro.",
+    ],
+    liked: false,
+    destacado: true,
+    user: "Chapulin Colorado",
+    userProfile: "https://pbs.twimg.com/media/EVcjT-JXQAIq0Yy.jpg",
+  },
+];
+
+export default content;
